Remove ios example and old cordova script
This removes the ios platform target for now, as it fails on mac osx due
to a child_process error that I haven't tracked down yet. This doesn't
setup the config.xml to include the icon and splash image values yet.
I'm still thinking about the best way to do that. Having to configure
every size in the package.json feels cumbersome. I like the idea of
configuring image root folders for icon/splash per platform and the
script adding the config nodes in the config.xml to point to expected
file names. The docs could then say what images are expected for each
platform and the devlopers would just put those images at those sizes
into the respective folders for packaging. This is how Ionic does it,
though they provide a CLI tool to generate the images as well.